+++ dataland-frontend/tests/e2e/support/index.ts
@@ -1,2 +1,2 @@
 import '@cypress/code-coverage/support'
-import "./command"
+import "./commands"
